src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Jonathan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.118.2"}
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
pydantic-settings = "^2.11.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.10.1"
ruff = "^0.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120 # PEP-8

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["PT001", "ANN201", "D103", "S101", "ANN001", "PT004", "ARG001"]
"**/schemas/*" = ["D101"]
"**/settings.py" = ["D101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
ignore = ["D407", "D100", "FA102", "D104", "PLC0414", "INP001", "PTH123"]
select = ['ALL']
extend-select = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.taskipy.tasks]
test = 'python -m pytest -s --cov=. -x -vv --cov-config=.coveragerc'
dev = 'fastapi dev main.py --host 0.0.0.0 --port 8001'
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
